split middleware to a separate file
